+++ lib/ui/gbv/GBVQuickExitWidget.dart
@@ -25,7 +25,7 @@ class GBVQuickExitWidget extends StatelessWidget {
         TextSpan(children: [
           TextSpan(text: Localization().getStringEx('widget.sexual_misconduct.quick_exit.privacy', 'Privacy: '),
               style: Styles().textStyles.getTextStyle('widget.item.small.fat')),
-          TextSpan(text: Localization().getStringEx('widget.sexual_misconduct.quick_exit.app_activity', 'your app activity is not shared with others.'),
+          TextSpan(text: Localization().getStringEx('widget.sexual_misconduct.quick_exit.app_activity', 'your personal app activity is not shared with others.'),
               style: Styles().textStyles.getTextStyle('widget.item.small.thin')),
         ])
         )